TOLP-5939 Some columns need BigInteger

+++ tolqc-api/app/main/model/data.py
@@ -36,7 +36,7 @@ class Data(LogBase):
     withdrawn = db.Column(db.Boolean())
     manually_withdrawn = db.Column(db.Boolean())
     reads = db.Column(db.Integer())
-    bases = db.Column(db.Integer())
+    bases = db.Column(db.BigInteger())
     read_length_mean = db.Column(db.Float())
     read_length_n50 = db.Column(db.Integer())
     elastic_mlwh_cksum = db.Column(db.String())
